+++ pyportable_installer/__init__.py
@@ -1,4 +1,4 @@
 from .main import debug_build, full_build, min_build
 # from .venv_builder import download_embed_python
 
-__version__ = '3.2.2'
+__version__ = '3.3.0'