+++ pyportable_installer/main.py
@@ -5,6 +5,7 @@ from lk_logger import lk
 from .no1_extract_pyproject import main as step1
 from .no2_prebuild_pyproject import main as step2
 from .no3_build_pyproject import main as step3
+from .typehint import TConf
 
 
 class Misc:
@@ -20,6 +21,9 @@ class Misc:
     compile_scripts = True
     # 是否做善后工作. 如是, 详见 `aftermath.py` 模块.
     do_aftermath = True
+    # 是否打印详细信息. 默认是 False (关闭). 注意即使是 False, 也会打印原本应该
+    # 打印的内容, 只是 False 状态不会打印 sourcemap 信息.
+    log_verbose = False
     
     @classmethod
     def dump(cls):
@@ -29,6 +33,7 @@ class Misc:
             'create_launch_bat'   : cls.create_launch_bat,
             'compile_scripts'     : cls.compile_scripts,
             'do_aftermath'        : cls.do_aftermath,
+            'log_verbose'         : cls.log_verbose,
         }
 
 
@@ -36,14 +41,14 @@ def full_build(pyproj_file):
     Misc.create_checkup_tools = True
     Misc.create_venv_shell = True
     Misc.create_launch_bat = True
-    main(pyproj_file, Misc.dump())
+    return main(pyproj_file, Misc.dump())
 
 
 def min_build(file):
     Misc.create_checkup_tools = False
     Misc.create_venv_shell = False
     Misc.create_launch_bat = True  # suggest True
-    main(file, Misc.dump())
+    return main(file, Misc.dump())
 
 
 def debug_build(file):
@@ -52,15 +57,19 @@ def debug_build(file):
     Misc.create_launch_bat = True
     Misc.compile_scripts = False
     Misc.do_aftermath = False
-    main(file, Misc.dump())
+    Misc.log_verbose = True
+    return main(file, Misc.dump())
 
 
-def main(pyproj_file: str, misc: dict):
+def main(pyproj_file: str, misc: dict) -> TConf:
     """
     几个关键目录的区分和说明: `docs/devnote/difference-between-roots.md`
     """
+    if misc.get('log_verbose', False) is False:
+        lk.lite_mode = True
+    
     prj_conf = step1(pyproj_file)
-    ________ = step2(prj_conf)  # 这里用下划线作变量, 只是为了对齐, 使代码美观
+    ________ = step2(prj_conf)  # 下划线仅用于对齐代码, 无特别作用
     dst_root = step3(prj_conf['app_name'], **prj_conf['build'], **misc)
     
     m, n = ospath.split(dst_root)
@@ -70,3 +79,5 @@ def main(pyproj_file: str, misc: dict):
     if misc.get('do_aftermath', True):
         from .aftermath import main as do_aftermath
         do_aftermath(pyproj_file, prj_conf, dst_root)
+    
+    return prj_conf